requirements.txt
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oracledb==3.4.2
pycparser==3.0
python-dotenv==1.2.2
typing_extensions==4.15.0
Werkzeug==3.1.6
